backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
celery[redis]==5.3.6
redis==5.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
python-multipart==0.0.9
pydantic-settings==2.2.1
python-docx==1.1.2
PyMuPDF==1.24.3
